Integrate getContent into hooks function, Alert component for showing error messages
